+++ ouster-ros/ouster-sdk/docs/_ext/init.py
@@ -1,3 +1,0 @@
-"""
-Internal documentation build helpers for the Ouster SDK.
-"""