+++ ouster-ros/ouster-sdk/python/src/ouster/cli/__init__.py
@@ -1,2 +1,0 @@
-from pkgutil import extend_path
-__path__ = extend_path(__path__, __name__)  # Declare as a namespace package - see PEP 420