+++ ouster-ros/ouster-sdk/python/src/ouster/sdk/examples/__init__.py
@@ -1,9 +1,0 @@
-"""
-Copyright (c) 2021, Ouster, Inc.
-All rights reserved.
-
-Code examples for using the Ouster SDK.
-
-These modules are provided for documentation and testing purposes only, and
-should not be considered a stable public API.
-"""